cloud_dataflow/requirements.txt
apache-beam[gcp]==2.53.0
pandas==2.1.4
pandas-gbq==0.20.0
google-cloud-bigquery==3.10.0
google-cloud-storage==2.16.0
gcsfs
fsspec==2024.2.0